Added two types of message: LoginOutcome and LoginResponse; Changed client and server, added saved accounts.
Changed how client and server interact: now both of them handle messages in a modular way.
Now a username and a password are required to the client to access the server, if the credentials provided are not saved yet, the server proceeds to create a new account.
All the credentials are saved in an account.txt file with a serialized HashMap, if the file is not present, the server proceeds to create a new one.