Fixed Login, Join, Lobby and LoadSave procedures.
Now there can be only one client with the same credentials connected per time.
If the Accounts.ser file doesn't exist, the MainServer proceeds to create a new one.
If there are no longer players in the Lobby, LobbiesHandler proceeds to delete the Lobby.
If a client disconnects from the server while joined a Lobby, it gets removed from the Lobby.