GUI updated
Now the GUI is updated to show the board to the players, turns implementation still missing and interface needs to be finished.
N.B. if you want to see the images you have to put the folder containing them in your local version of the project in the "resources" directory.